Refs #4993 Moved KeyedChanges to a common/ subdirectory
